feat: entity deletion actions

+++ src/comments/comment.service.ts
@@ -1,9 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Comment } from './comment.entity';
 import { CreateCommentInput } from './dto/comment.dto';
 import { TicketService } from '~/tickets/ticket.service';
+import authErrors from '~/auth/const/auth.errors';
+import ticketErrors from '~/tickets/ticket.constants';
 
 @Injectable()
 export class CommentService {
@@ -20,6 +26,7 @@ export class CommentService {
       where: {
         ticket: { id: ticketId },
       },
+      relations: ['author'],
     });
   }
 
@@ -34,7 +41,20 @@ export class CommentService {
     });
   }
 
-  delete(id: string) {
+  async delete(userId: string, id: string) {
+    const comment = await this.commentRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
+
+    if (!comment) {
+      throw new NotFoundException(ticketErrors.COMMENT_NOT_FOUND);
+    }
+
+    if (comment.author.id !== userId) {
+      throw new ForbiddenException(authErrors.ROLE_MISMATCH);
+    }
+
     return this.commentRepository.delete(id);
   }
 }
